move code hashing helpers to core
